Refactored and finished up code coverage report, by commenting on PR
