+++ src/index.ts
@@ -1,27 +1,18 @@
-import table from "markdown-table";
-import path from "path";
-
-import { getInput } from "@actions/core";
 import { issueCommand, issue } from "@actions/core/lib/command";
-import { context, getOctokit } from "@actions/github";
-import { GitHub } from "@actions/github/lib/utils";
+import { postCodeCoverage } from "./code-coverage";
+import { symlinkSync } from "fs";
 
 interface GitHubActionsReporterOptions {
     relativeDirectories?: boolean;
+    postCodeCoverageComment?: boolean;
 }
-
-type File = {
-    relative: string;
-    fileName: string;
-    path: string;
-    coverage: jest.CoverageSummary;
-};
 
 class GitHubActionsReporter implements jest.Reporter {
     private regex = /\((.+?):(\d+):(\d+)\)/;
 
     private options: GitHubActionsReporterOptions = {
         relativeDirectories: false,
+        postCodeCoverageComment: false,
     };
 
     constructor(public globalConfig: jest.GlobalConfig, options: GitHubActionsReporterOptions) {
@@ -47,10 +38,21 @@ class GitHubActionsReporter implements jest.Reporter {
 
         issue("endgroup");
 
-        if (result.coverageMap) {
+        if (this.options.postCodeCoverageComment) {
+            if (!result.coverageMap) {
+                console.error(
+                    "jest-github-actions-reporter was instructed to post code coverage comment, but code coverage is not enabled in jest. \n" +
+                        "Set collectCoverage to true or postCodeCoverageComment to false."
+                );
+                process.exit(50);
+            }
+
             console.log("Posting code coverage results as comment");
 
-            this.postCodeCoverage(result.coverageMap);
+            postCodeCoverage(result.coverageMap).catch((err: Error) => {
+                console.error(err.message);
+                process.exit(51);
+            });
         }
     }
 
@@ -71,69 +73,6 @@ class GitHubActionsReporter implements jest.Reporter {
                 issueCommand("error", args, failureMessage);
             }
         }
-    }
-
-    private async postCodeCoverage(coverageMap: jest.CoverageMap) {
-        const githubToken = getInput("github-token");
-        const octokit = getOctokit(githubToken);
-
-        const t = this.generateCoverageTable(coverageMap);
-
-        await octokit.issues.createComment({
-            repo: context.repo.repo,
-            owner: context.repo.owner,
-            body: t,
-            issue_number: context.payload.number,
-        });
-    }
-
-    private generateCoverageTable(coverageMap: jest.CoverageMap) {
-        const formatIfPoor = (number: number, threshold = 50): string => {
-            return number < threshold ? `${number} :red_circle:` : `${number} :green_circle:`;
-        };
-
-        const summaryToRow = (f: jest.CoverageSummary) => [
-            formatIfPoor(f.statements.pct!),
-            formatIfPoor(f.branches.pct!),
-            formatIfPoor(f.functions.pct!),
-            formatIfPoor(f.lines.pct!),
-        ];
-
-        const parseFile = (absolute: string) => {
-            const relative = path.relative(process.cwd(), absolute);
-            const fileName = path.basename(relative);
-            const p = path.dirname(relative);
-            const coverage = coverageMap.fileCoverageFor(absolute).toSummary();
-            return { relative, fileName, path: p, coverage };
-        };
-
-        const groupByPath = (dirs: { [key: string]: File[] }, file: File) => {
-            if (!(file.path in dirs)) {
-                dirs[file.path] = [];
-            }
-
-            dirs[file.path].push(file);
-
-            return dirs;
-        };
-
-        const header = ["File", "% Statements", "% Branch", "% Funcs", "% Lines"];
-        const summary = (coverageMap.getCoverageSummary() as unknown) as jest.CoverageSummary;
-        const summaryRow = ["**All**", ...summaryToRow(summary)];
-
-        const files = coverageMap.files().map(parseFile).reduce(groupByPath, {});
-
-        const rows = Object.entries(files)
-            .map(([dir, files]) => [
-                [` **${dir}**`, "", "", "", ""], // Add metrics for directories by summing files
-                ...files.map((file) => {
-                    const name = `\`${file.fileName}\``;
-                    return [`  ${name}`, ...summaryToRow(file.coverage)];
-                }),
-            ])
-            .flat();
-
-        return table([header, summaryRow, ...rows], { align: ["l", "r", "r", "r", "r"] });
     }
 }
